requirements.txt
django==2.0.*
django-environ==0.4.*
dj-database-url==0.5.*
drf_yasg==1.11.*
django_extensions==2.1.*
django-choices==1.6.*
drf-nested-routers==0.91
djangorestframework==3.9.*
pillow==2.9.0
psycopg2-binary==2.7.*
whitenoise==3.3.*
django-cors-headers==2.4.*
django-background-tasks==1.1.*

django-coverage-plugin==1.6.*
django-debug-toolbar==1.11
factory-boy==2.12.*
Faker==1.0.*
flake8==3.7.*
pycodestyle==2.5.*
pytest==3.10.1
pytest-cov==2.6.*
pytest-xdist==1.28.*
pytest-django==3.4.*
pytest-factoryboy==2.0.*
pytest-mock==1.10.*
pytest-sugar==0.9.*
requests-mock==1.6.*

gevent
gunicorn
newrelic